# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Радар обязанностей — API мини-приложения",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/dashboard": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Главная: открытые задачи, счётчики, ближайший срок, есть ли непрочитанное",
        "operationId": "dashboard_api_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardData"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/calendar": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Задачи со сроком в интервале дат from–to",
        "operationId": "calendar_api_calendar_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "title": "Response Calendar Api Calendar Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feed": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Лента: о чём бот писал в чат за 90 дней и чем кончилось",
        "operationId": "feed_api_feed_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedItem"
                  },
                  "type": "array",
                  "title": "Response Feed Api Feed Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Экран события: почему вам, что сделать, куда сдавать, основание, штраф",
        "operationId": "task_details_api_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/submitted": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "«Выполнено»: задача закрыта, пуш о ней уходит из чата",
        "operationId": "mark_submitted_api_tasks__task_id__submitted_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Отменить «Выполнено»",
        "operationId": "undo_submitted_api_tasks__task_id__submitted_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/list": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "«В список дел»: событие на главной, пуш о нём уходит из чата",
        "operationId": "add_to_list_api_tasks__task_id__list_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/document": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Подготовить черновик документа (.docx) — бот пришлёт его в чат; 404 — у задачи нет документа, 502 — MAX не принял сообщение",
        "operationId": "prepare_document_api_tasks__task_id__document_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Counters": {
        "properties": {
          "overdue": {
            "type": "integer",
            "title": "Overdue"
          },
          "soon": {
            "type": "integer",
            "title": "Soon"
          },
          "done": {
            "type": "integer",
            "title": "Done"
          }
        },
        "type": "object",
        "required": [
          "overdue",
          "soon",
          "done"
        ],
        "title": "Counters"
      },
      "DashboardData": {
        "properties": {
          "counters": {
            "$ref": "#/components/schemas/Counters"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "type": "array",
            "title": "Tasks"
          },
          "nextDue": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextdue"
          },
          "unread": {
            "type": "boolean",
            "title": "Unread"
          }
        },
        "type": "object",
        "required": [
          "counters",
          "tasks",
          "unread"
        ],
        "title": "DashboardData"
      },
      "FeedItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "subtitle": {
            "type": "string",
            "title": "Subtitle"
          },
          "sentAt": {
            "type": "string",
            "format": "date-time",
            "title": "Sentat"
          },
          "due": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "done",
              "muted"
            ],
            "title": "Status"
          },
          "listed": {
            "type": "boolean",
            "title": "Listed"
          },
          "severity": {
            "type": "string",
            "enum": [
              "critical",
              "warning",
              "info"
            ],
            "title": "Severity"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "subtitle",
          "sentAt",
          "status",
          "listed",
          "severity"
        ],
        "title": "FeedItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Link": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "label",
          "url"
        ],
        "title": "Link"
      },
...
      "Task": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "subtitle": {
            "type": "string",
            "title": "Subtitle"
          },
          "status": {
            "type": "string",
            "enum": [
              "overdue",
              "soon",
              "planned",
              "done"
            ],
            "title": "Status"
          },
          "due": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due"
          },
          "periodicity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodicity"
          },
          "listed": {
            "type": "boolean",
            "title": "Listed",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "subtitle",
          "status"
        ],
        "title": "Task"
      },
      "TaskDetails": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "subtitle": {
            "type": "string",
            "title": "Subtitle"
          },
          "status": {
            "type": "string",
            "enum": [
              "overdue",
              "soon",
              "planned",
              "done"
            ],
            "title": "Status"
          },
          "due": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due"
          },
          "periodicity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodicity"
          },
          "listed": {
            "type": "boolean",
            "title": "Listed",
            "default": false
          },
          "heading": {
            "type": "string",
            "title": "Heading"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/TaskSection"
            },
            "type": "array",
            "title": "Sections"
          },
          "document": {
            "type": "boolean",
            "title": "Document",
            "default": false
          },
          "next": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Task"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "subtitle",
          "status",
          "heading",
          "sections"
        ],
        "title": "TaskDetails"
      },
      "TaskSection": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "caption": {
            "type": "string",
            "title": "Caption"
          },
          "body": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "steps": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Steps"
          },
          "link": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Link"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "icon",
          "title",
          "caption"
        ],
        "title": "TaskSection"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}